Refactor: 열린, 닫힌 이슈 탭 클릭시 불필요한 리렌더링이 발생하여 리팩토링
