+++ FE/issue-tracker/src/components/issues/tableHeader/TableHeader.tsx
@@ -1,41 +1,15 @@
 import styled from 'styled-components';
-import { useRecoilValueLoadable } from 'recoil';
-
-import { issueCounts } from '@store/atoms/issueList';
 
 import AssigneeFilter from './AssigneeFilter';
 import LabelFilter from './LabelFilter';
 import MilestoneFilter from './MilestoneFilter';
 import AuthorFilter from './AuthorFilter';
+import TableHeadLeft from './TableHeadLeft';
 
-function TableHeader({ isClosed, handleClickTab }: Props) {
-  const { state, contents } = useRecoilValueLoadable(issueCounts);
-
+function TableHeader() {
   return (
     <TableHeaderWrap>
-      <HeaderLeft>
-        <input type="checkbox" name="" id="" />
-        <IssueTab>
-          <IssueOpenTab
-            isClosed={isClosed}
-            data-state="open"
-            onClick={handleClickTab}
-          >
-            {state === 'loading' && `열린 이슈 (0)`}
-            {state === 'hasValue' && `열린 이슈 (${contents.openIssueCount})`}
-            {state === 'hasError' && `열린 이슈 (??)`}
-          </IssueOpenTab>
-          <IssueCloseTab
-            isClosed={isClosed}
-            data-state="close"
-            onClick={handleClickTab}
-          >
-            {state === 'loading' && `닫힌 이슈 (0)`}
-            {state === 'hasValue' && `닫힌 이슈 (${contents.closeIssueCount})`}
-            {state === 'hasError' && `닫힌 이슈 (??)`}
-          </IssueCloseTab>
-        </IssueTab>
-      </HeaderLeft>
+      <TableHeadLeft />
       <HeaderRight>
         <FilterLists>
           <AuthorFilter />
@@ -49,15 +23,6 @@ function TableHeader({ isClosed, handleClickTab }: Props) {
 }
 
 export default TableHeader;
-
-type Props = {
-  handleClickTab: (e: React.MouseEvent) => void;
-  isClosed: boolean;
-};
-
-type issueTab = {
-  isClosed: boolean;
-};
 
 const TableHeaderWrap = styled.div`
   margin-top: 24px;
@@ -89,27 +54,4 @@ const HeaderRight = styled.div`
 
 const FilterLists = styled.div`
   display: flex;
-`;
-
-const IssueTab = styled.ul`
-  display: flex;
-`;
-
-const IssueTabList = styled.li<issueTab>`
-  ${({ theme }) => theme.flexCenter}
-  margin-left: 24px;
-  width: 100px;
-  height: 28px;
-  cursor: pointer;
-  font-weight: ${({ theme }) => theme.fontWeights.bold};
-`;
-
-const IssueOpenTab = styled(IssueTabList)`
-  color: ${({ theme, isClosed }) =>
-    isClosed ? theme.colors.gr_label : theme.colors.gr_titleActive};
-`;
-
-const IssueCloseTab = styled(IssueTabList)`
-  color: ${({ theme, isClosed }) =>
-    isClosed ? theme.colors.gr_titleActive : theme.colors.gr_label};
 `;
